requirements.txt
adal==1.2.1
aiohttp==3.6.2
aiounittest==1.3.1
async-timeout==3.0.1
attrs==19.3.0
azure-cognitiveservices-language-luis==0.2.0
azure-common==1.1.25
botbuilder-ai==4.7.1
botbuilder-core==4.7.1
botbuilder-schema==4.7.1
botframework-connector==4.7.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.8
Flask==1.1.1
idna==2.8
idna-ssl==1.1.0
importlib-metadata==1.5.0
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonpickle==1.2
MarkupSafe==1.1.1
more-itertools==8.2.0
msrest==0.6.10
msrestazure==0.6.2
multidict==4.7.5
oauthlib==3.1.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
pycparser==2.20
PyJWT==1.5.3
pyparsing==2.4.6
pytest==5.3.5
pytest-aiohttp==0.3.0
python-dateutil==2.8.1
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
typing-extensions==3.7.4.1
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==1.0.0
yarl==1.4.2
zipp==3.1.0
